moved get/set/dump/dirty to base config class instead of root
